Revert "fix: Make `SearchHistoryId` parcelable"
This reverts commit 57896c27282bfbc075afb3f73bdd86b04b1a6ea5.

+++ app/src/main/kotlin/com/prajwalch/torrentsearch/data/local/dao/SearchHistoryDao.kt
@@ -18,7 +18,7 @@ interface SearchHistoryDao {
     fun observeAll(): Flow<List<SearchHistoryEntity>>
 
     @Query("DElETE from search_history where id=:id")
-    suspend fun deleteById(id: Long)
+    suspend fun deleteById(id: Int)
 
     @Query("DELETE from search_history")
     suspend fun deleteAll()
